+++ Scripts/ParseEmailFiles/ParseEmailFiles.py
@@ -191,8 +191,9 @@ class DataModel(object):
     def PtypString(data_value):
         if data_value:
             try:
-                encoding = chardet.detect(data_value)
-                data_value = data_value.decode(encoding['encoding'], errors='ignore').replace('\x00', '')
+                res = chardet.detect(data_value)
+                enc = res['encoding'] or 'ascii'  # in rare cases chardet fails to detect and return None as encoding
+                data_value = data_value.decode(enc, errors='ignore').replace('\x00', '')
             except UnicodeDecodeError:
                 data_value = data_value.decode("utf-16-le", errors="ignore").replace('\x00', '')
 
@@ -407,11 +408,11 @@ class EmailFormatter(object):
             if maintype == 'text' or "message" in maintype:
                 attach = MIMEText(data, _subtype=subtype)
             elif maintype == 'image':
-                attach = MIMEImage(data, _subtype=subtype)  # type: ignore
+                attach = MIMEImage(data, _subtype=subtype)  # type: ignore[assignment]
             elif maintype == 'audio':
-                attach = MIMEAudio(data, _subtype=subtype)  # type: ignore
+                attach = MIMEAudio(data, _subtype=subtype)  # type: ignore[assignment]
             else:
-                attach = MIMEBase(maintype, subtype)  # type: ignore
+                attach = MIMEBase(maintype, subtype)  # type: ignore[assignment]
                 attach.set_payload(data)
 
                 # Encode the payload using Base64
@@ -2932,7 +2933,7 @@ class Message(object):
         if property_value:
             property_detail = {property_name: property_value}
         else:
-            property_detail = None  # type: ignore
+            property_detail = None  # type: ignore[assignment]
 
         return property_detail
 
@@ -3158,7 +3159,7 @@ def parse_email_headers(header, raw=False):
     if raw:
         return headers
 
-    email_address_headers = {  # type: ignore
+    email_address_headers = {  # type: ignore[var-annotated]
         "To": [],
         "From": [],
         "CC": [],
@@ -3292,8 +3293,8 @@ def save_attachments(attachments, root_email_file_name, max_depth):
         if attachment.data is not None:
             display_name = attachment.DisplayName if attachment.DisplayName else attachment.AttachFilename
             demisto.results(fileResult(display_name, attachment.data))
-
-            if max_depth > 0 and display_name.lower().endswith(".eml"):
+            name_lower = display_name.lower()
+            if max_depth > 0 and (name_lower.endswith(".eml") or name_lower.endswith('.p7m')):
                 tf = tempfile.NamedTemporaryFile(delete=False)
 
                 try:
@@ -3302,10 +3303,12 @@ def save_attachments(attachments, root_email_file_name, max_depth):
 
                     inner_eml, attached_inner_emails = handle_eml(tf.name, file_name=root_email_file_name,
                                                                   max_depth=max_depth)
-                    return_outputs(readable_output=data_to_md(inner_eml, attachment.DisplayName, root_email_file_name),
-                                   outputs=None)
-                    attached_emls.append(inner_eml)
-                    attached_emls.extend(attached_inner_emails)
+                    if inner_eml:
+                        return_outputs(readable_output=data_to_md(inner_eml, attachment.DisplayName, root_email_file_name),
+                                       outputs=None)
+                        attached_emls.append(inner_eml)
+                    if attached_inner_emails:
+                        attached_emls.extend(attached_inner_emails)
                 finally:
                     os.remove(tf.name)