+++ Integrations/AutofocusV2/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [19.10.0] - 2019-10-03
+  - Improved handling of empty responses for the  ***autofocus-samples-search*** and ***autofocus-sessions-search*** commands.
+
+
 ## [19.9.1] - 2019-09-18
 Added several arguments to the ***autofocus-samples-search*** and ***autofocus-sessions-search*** commands.
   - *file_hash*