+++ Integrations/SentinelOne-V2/CHANGELOG.md
@@ -1,6 +1,15 @@
 ## [Unreleased]
 
 
+## [19.10.2] - 2019-10-29
+  - Fixed an issue in the ***Fetch incidents*** function.
+  - Fixed an issue in the ***sentinelone-get-threats*** command.
+
+
+## [19.10.0] - 2019-10-03
+  - Fixed an issue with the ***sentinelone-disconnect-agent*** command.
+  - Fixed human-readable output in the ***sentinelone-get-threat*** command in cases where the content_hash does not exist.
+
 ## [19.9.1] - 2019-09-18
   - Fixed an issue in the **Fetch incidents*** function.
   - Fixed date parameters in the ***sentinelone-get-threats*** command.